+++ src/Microsoft.ApplicationInsights.AspNetCore/Implementation/TelemetryModuleConfigurator.cs
@@ -1,6 +1,7 @@
 ﻿namespace Microsoft.ApplicationInsights.AspNetCore
 {
     using System;
+    using Microsoft.ApplicationInsights.AspNetCore.Extensions;
     using Microsoft.ApplicationInsights.Extensibility;
 
     /// <summary>
@@ -8,28 +9,31 @@
     /// </summary>
     internal class TelemetryModuleConfigurator : ITelemetryModuleConfigurator
     {
-        private readonly Action<ITelemetryModule> configure;
-        internal readonly Type telemetryModuleType;
+        private readonly Action<ITelemetryModule, ApplicationInsightsServiceOptions> configure;
 
         /// <summary>
         /// Constructs an instance of <see cref="TelemetryModuleConfigurator"/>.
         /// </summary>        
         /// <param name="configure">The action used to configure telemetry module.</param>
         /// <param name="telemetryModuleType">The type of the telemetry module being configured.</param>
-        public TelemetryModuleConfigurator(Action<ITelemetryModule> configure, Type telemetryModuleType)
+        public TelemetryModuleConfigurator(Action<ITelemetryModule, ApplicationInsightsServiceOptions> configure, Type telemetryModuleType)
         {
             this.configure = configure;
             this.TelemetryModuleType = telemetryModuleType;
         }
 
-        /// <summary>
-        /// Creates an instance of the telemetry processor, passing the
-        /// next <see cref="ITelemetryProcessor"/> in the call chain to
-        /// its constructor.
-        /// </summary>
+        [Obsolete("Use Configure(ITelemetryModule telemetryModule, ApplicationInsightsServiceOptions options) instead.", true)]
         public void Configure(ITelemetryModule telemetryModule)
         {
-            this.configure?.Invoke(telemetryModule);
+            this.configure?.Invoke(telemetryModule, null);
+        }
+
+        /// <summary>
+        /// Configures telemetry module.
+        /// </summary>
+        public void Configure(ITelemetryModule telemetryModule, ApplicationInsightsServiceOptions options)
+        {
+            this.configure?.Invoke(telemetryModule, options);
         }
 
         /// <summary>