+++ test/MVCFramework20.FunctionalTests/FunctionalTest/ExceptionTelemetryMvcTests.cs
@@ -29,7 +29,9 @@ namespace MVCFramework20.FunctionalTests.FunctionalTest
                 expectedRequestTelemetry.ResponseCode = "500";
                 expectedRequestTelemetry.Success = false;
                 expectedRequestTelemetry.Url = new System.Uri(server.BaseHost + RequestPath);
-                this.ValidateBasicRequest(server, "/Home/Exception", expectedRequestTelemetry);
+
+                // the is no response header because of https://github.com/Microsoft/ApplicationInsights-aspnetcore/issues/717
+                this.ValidateBasicRequest(server, "/Home/Exception", expectedRequestTelemetry, false);
             }
         }